Add C-style array sizing functions
Dealing with C-style arrays can be dangerous. The semantics of the
sizeof operator is different for arrays and pointers but arrays can
decay into pointers without warning, especially across function calls
where they are automatically converted.

The functions provided allow retrieving the number of elements, total
size and element size of a C-style array in a safe fashion. Each will
cause a compile error if a pointer or other incompatible type is
provided.